Reorganized all tests into __test__, changed main workflow to run E2E and Unit tests separately with different env, modified jest configurations to support this, modified Coverage report generation to accomodate separated tests
