+++ __tests__/unit/unit.test.js
@@ -1,6 +1,6 @@
 import {
     dummy
-  } from '../functions.js';
+  } from '../../functions.js';
